Resolves #65270: Security issue in sr_feuser_register 3.1.3
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/sr_feuser_register@90360 735d13b6-9817-0410-8766-e36946ffe9aa

+++ trunk/Classes/Controller/DeleteActionController.php
@@ -60,13 +60,15 @@ class DeleteActionController extends AbstractActionController
 		if ($this->conf[$cmdKey . '.']['preview'] && (int) $this->parameters->getFeUserData('preview')) {
 			$mode = AbstractView::MODE_PREVIEW;
 		}
+		$isSubmit = $this->parameters->getFeUserData('submit');
+		$isSubmit = !empty($isSubmit);		
 		$isDoNotSave = $this->parameters->getFeUserData('doNotSave');
 		$isDoNotSave = !empty($isDoNotSave);
 		if ($isDoNotSave) {
 			SessionData::clearSessionData($this->extensionKey);
 		}
 		// If data is submitted, we take care of it here.
-		if ($mode === AbstractView::MODE_NORMAL && !$isDoNotSave) {
+		if ($mode === AbstractView::MODE_NORMAL && $isSubmit && !$isDoNotSave) {
 			$this->data->deleteRecord($origArray, $dataArray);
 			if ($this->data->getSaved()) {
 				$key =  'DELETE' . Marker::SAVED_SUFFIX;
